refactor(racing): 불필요하게 겹친 코드 제거
