+++ src/main/java/racing/domain/CarStatement.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public class CarStatement {
     private final String name;
