+++ src/test/java/racing/CarStatementTest.java
@@ -2,6 +2,7 @@ package racing;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racing.domain.CarStatement;
 
 import static org.assertj.core.api.Assertions.assertThat;
